Add latest lab solve.\n

+++ 06-02-2024/problem2.1.c
@@ -5,10 +5,10 @@ int main() {
   printf("Input your height in centimeter: ");
   scanf("%f", &centimeter);
 
-  int inch = centimeter * 0.394;
-  int feet = (inch / 12);
-  inch = inch % 12;
+  float inch = centimeter * 0.394;
+  int feet = inch / 12;
+  inch = (int)inch % 12;
 
-  printf("Your Height is %d feet and %d inch\n", feet, inch);
+  printf("Your Height is %d feet and %.0f inch\n", feet, inch);
   return 0;
 }
